It works now!
FossilOrigin-Name: fe4f5d35aa384de2617454d942b79ee5a2a13318c92050b05b9ad2639ca45dcf

+++ HealthMonitorLogViewer/StatsHourly.cs
@@ -9,11 +9,11 @@ public partial class StatsHourly
 
     public string ProcessName { get; set; }
 
-    public int? AvgWorkingSet { get; set; }
+    public long? AvgWorkingSet { get; set; }
 
     public string MaxWorkingSetForOneInstance { get; set; }
 
-    public int? CpuDiff { get; set; }
+    public long? CpuDiff { get; set; }
 
     public byte[] CpuPercent { get; set; }
